# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrailDifficulty = typeof TrailDifficulty[keyof typeof TrailDifficulty];


export const TrailDifficulty = {
  easy: 'easy',
  moderate: 'moderate',
  hard: 'hard',
  expert: 'expert',
} as const;

export interface Trail {
  id: number;
  name: string;
  difficulty: TrailDifficulty;
  distanceMiles: number;
  /** @nullable */
  elevationFt?: number | null;
  location: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type TrailInputDifficulty = typeof TrailInputDifficulty[keyof typeof TrailInputDifficulty];


export const TrailInputDifficulty = {
  easy: 'easy',
  moderate: 'moderate',
  hard: 'hard',
  expert: 'expert',
} as const;

export interface TrailInput {
  name: string;
  difficulty: TrailInputDifficulty;
  distanceMiles: number;
  /** @nullable */
  elevationFt?: number | null;
  location: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
}

export type TrailUpdateDifficulty = typeof TrailUpdateDifficulty[keyof typeof TrailUpdateDifficulty];


export const TrailUpdateDifficulty = {
  easy: 'easy',
  moderate: 'moderate',
  hard: 'hard',
  expert: 'expert',
} as const;

export interface TrailUpdate {
  name?: string;
  difficulty?: TrailUpdateDifficulty;
  distanceMiles?: number;
  /** @nullable */
  elevationFt?: number | null;
  location?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Event {
  id: number;
  title: string;
  date: string;
  location: string;
  description: string;
  /** @nullable */
  trailId?: number | null;
  /** @nullable */
  trailName?: string | null;
  attendeeCount: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface EventInput {
  title: string;
  date: string;
  location: string;
  description: string;
  /** @nullable */
  trailId?: number | null;
  attendeeCount?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface EventUpdate {
  title?: string;
  date?: string;
  location?: string;
  description?: string;
  /** @nullable */
  trailId?: number | null;
  attendeeCount?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Member {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  role?: string | null;
  trailsCompleted?: number;
  joinedAt: string;
}

export interface MemberInput {
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  role?: string | null;
  trailsCompleted?: number;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  pinned?: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  title: string;
  content: string;
  pinned?: boolean;
}

export type DashboardSummaryTrailsByDifficulty = {
  easy: number;
  moderate: number;
  hard: number;
  expert: number;
};

export interface DashboardSummary {
  totalTrails: number;
  totalMembers: number;
  upcomingEvents: number;
  totalDistanceMiles?: number;
  trailsByDifficulty: DashboardSummaryTrailsByDifficulty;
  recentAnnouncements: Announcement[];
}

export type ListTrailsParams = {
difficulty?: string;
};