# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  content: string;
  pinned?: boolean;
  createdAt: string;
}