# lib/api-zod/src/generated/types/announcementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementInput {
  title: string;
  content: string;
  pinned?: boolean;
}