# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */
import type { Announcement } from './announcement';
import type { DashboardSummaryTrailsByDifficulty } from './dashboardSummaryTrailsByDifficulty';

export interface DashboardSummary {
  totalTrails: number;
  totalMembers: number;
  upcomingEvents: number;
  totalDistanceMiles?: number;
  trailsByDifficulty: DashboardSummaryTrailsByDifficulty;
  recentAnnouncements: Announcement[];
}