# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  title: string;
  date: string;
  location: string;
  description: string;
  /** @nullable */
  trailId?: number | null;
  attendeeCount?: number;
  /** @nullable */
  imageUrl?: string | null;
}