# lib/api-zod/src/generated/types/trail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */
import type { TrailDifficulty } from './trailDifficulty';

export interface Trail {
  id: number;
  name: string;
  difficulty: TrailDifficulty;
  distanceMiles: number;
  /** @nullable */
  elevationFt?: number | null;
  location: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}