# lib/api-zod/src/generated/types/trailInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */
import type { TrailInputDifficulty } from './trailInputDifficulty';

export interface TrailInput {
  name: string;
  difficulty: TrailInputDifficulty;
  distanceMiles: number;
  /** @nullable */
  elevationFt?: number | null;
  location: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
}