# lib/api-zod/src/generated/types/trailUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hiking Club API
 * OpenAPI spec version: 0.1.0
 */
import type { TrailUpdateDifficulty } from './trailUpdateDifficulty';

export interface TrailUpdate {
  name?: string;
  difficulty?: TrailUpdateDifficulty;
  distanceMiles?: number;
  /** @nullable */
  elevationFt?: number | null;
  location?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
}